requirements.txt
tensorflow
streamlit-option-menu
numpy
seaborn
matplotlib
scikit-learn
joblib
streamlit==1.26.0
cryptography
python-dotenv